+++ simulation/forces/internal_force.c
@@ -1,7 +1,7 @@
 #include "./internal_force.h"
 
 // force acting on the i-th body from the j-th body at the d-th dimention
-double fInternal(int i, int j, int d, double x[N][D], double m[N])
+long double fInternal(int i, int j, int d, long double x[N][D], long double m[N])
 {
     if (i == 0)
     // on centered body is acting no force
@@ -15,13 +15,13 @@ double fInternal(int i, int j, int d, double x[N][D], double m[N])
     }
     else
     {
-        double distance = 0;
+        long double distance = 0;
         for (int k = 0; k < D; k++)
         {
             distance += pow(x[i][k] - x[j][k], 2);
         }
 
-        double force = -1 * G * m[i] * m[j] / pow(distance, 3.0 / 2.0) * (x[i][d] - x[j][d]);
+        long double force = -1 * G * m[i] * m[j] / pow(distance, 3.0 / 2.0) * (x[i][d] - x[j][d]);
 
         return force;
     }
