+++ simulation/utilities/input_parameters.c
@@ -2,30 +2,30 @@
 #include "input_parameters.h"
 
 // Initialvelocities of all bodies
-double initial_velocities(int n, int d)
+long double initial_velocities(int n, int d)
 {
-	const double initial_velocities[N][D] = {{0, 0},
+	const long double initial_velocities[N][D] = {{0, 0},
 											 {325634, -23630},
 											 {15324, -36234}};
 	return initial_velocities[n][d];
 }
 
 // Mass of all bodies
-double initial_masses(int n)
+long double initial_masses(int n)
 {
-	const double initial_masses[N] = {542781632,
+	const long double initial_masses[N] = {542781632,
 									  12414421,
 									  125252};
 	return initial_masses[n];
 }
 
 // InitialIposition of all bodies (first entry = Saturn, second = moon 1, third = moon 2)
-double initial_pos(int b, int z)
+long double initial_pos(int b, int z)
 {
-	const double initial_positions[N][D] = {{0, 0},
+	const long double initial_positions[N][D] = {{0, 0},
 											{631952837, 6532892},
 											{4567843432, -467456634}};
 	return initial_positions[b][z];
 }
 
-double h = 0.05;
+long double h = 0.05;
